requirements.txt
fastapi[all]
uvicorn[standard]
sqlmodel
asyncpg
psycopg2-binary
pydantic-settings
python-dotenv
alembic
httpx
pytest
pytest-asyncio
aiofiles
passlib[bcrypt]
pyjwt
fastapi-mail
python-multipart
bcrypt==4.0.1
